requirements.txt
asgiref==3.8.1
azure-ai-documentintelligence==1.0.1
azure-core==1.32.0
azure-storage-blob==12.25.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
dnspython==2.7.0
idna==3.10
isodate==0.7.2
pillow==11.1.0
pycparser==2.22
pymongo==4.11.3
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
